Assume service team test role (#30)
Presubmit job wrapper will now attempt to look for an SSM param with the path `/ack/prow/service_team_role/$AWS_SERVICE`. If it finds a role by that name, it will assume it for the maximum amount of time of 1 hour (capped by chained assumption limitations - [see here](https://aws.amazon.com/premiumsupport/knowledge-center/iam-role-chaining-limit/)). This role is then used as the default role for the pod and for the KIND cluster's controller credentials.

Also included in this PR is documentation about onboarding with Prow and how to set up an AWS account in a manner compatible with the Prow job specifications.