Publish controller logs as Prow artifact (#71)
This pull request adds a new environment variable `$DUMP_CONTROLLER_LOGS` to the `kind-build-test.sh` script that optionally dumps the controller pod logs into the `$ARTIFACTS/controller_logs` folder. This folder is consumed by Prow and uploaded to S3 at the conclusion of the job.

It also configures `spyglass` with to add the `controller_logs` file to the `buildlog` lense, so that it is accessible from the web dashboard.