Fix Prometheus rules and allowlist (#119)
Description of changes:
The original document where I found the Prow dashboards and rules was for a much older version of `kube-state-metrics`. Some of the metrics it references in the rules have since been replaced by ones with different names. I have updated the names of these metrics, and fixed the syntax in the metrics allow list.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.