improve logging for condition wait (#126)
Often, a condition wait will fail during a test run and it's not straightforward to figure out why. Logging the resource state can help. For example, this could reveal that the wait for `ResourceSynced` to turn `True` failed due to a terminal condition.


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.