calculate PROPOSED and PLANNED project stages
Adds code to the ack-discover utility that calculates whether an AWS
service is in the PROPOSED or PLANNED project stage.

To determine the PROPOSED services, we look for a Github Issue with the
"Service Controller" label and a title matching either the abbreviated
or full service name.

To determine the PLANNED services, we check to see if the Github Issue
for a PROPOSED service is in the "Planned" Github ProjectColumn of the
"Service Controller Release Roadmap" Github Project in the community
repo.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>