Check code-gen version for auto-gen controllers (#196)
Description of changes:
When auto-generating controllers, don't immediately skip the controller if the runtime version does not match. Instead, check if there is a difference in the ack-generate binary version as compared to the generated metadata file.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.