Convert ack-discover to Python package (#198)
Description of changes:
In order to generate the service page for the documentation website, by being able to re-use the `ack-discover` CLI tool python methods, I converted it into a Python package with its own separate `setup.py` file. This can be installed independently from `acktest`, by installing it and using `import ackdiscover`.

I also made a few small changes to support the properties of controllers, releases and services we need for the website.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.