Add support for S3Bucket policy (#209)
Issue #, if available:

Description of changes:
To correctly execute cloudtrail-controller e2e we will need to set the
trai bucket policy. This patch adds possibility of creating a bucket and
setting it policy directly with the test-infra e2e test framework.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.