Fix Prow web identity file passing (#234)
Description of changes:
Fixes the authentication issues for the Prow cluster's test scripts. As well as a bug in the container credentials file, the file was not properly being passed to the Pytest container underneath. This PR was tested against the production Prow cluster to ensure it works for our IRSA setup.

Note: I have also removed the Prowjob's CPU requests, since there is no need for them to access 2 vCPUs worth of resources for every test run. This is unrelated to the web identity file work.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.