Upgrade to CDK 2 and EKS to 1.24 (#301)
Description of changes:
CDK 1 only supports EKS up to 1.21, which is now considered deprecated. In order to upgrade the cluster, I had to upgrade to CDK 2. Along with upgrading to 1.24, I had to upgrade the Prow control plane components that are now considered deprecated (namely Ingress `betav1` and CustomResourceDefinition `betav1`).

The new Prow CRD is 43k lines long, which is too big to be supported by Helm. To manage its installation, it is being included as a separate Kustomization, and then the Prow control plane, data plane and jobs are added to another Kustomization which has a dependency on the CRDs (so they are installed in the correct order). 

This system was tested by deploying it to my local AWS account, pointing the cluster to my private branch so that I could check that all of the Flux parts reconciled properly. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.