Make all code-generator pre-submit tests required.
Recently, code-generator PRs were being merged as soon as unit tests
passed, regardless of th eoutcome of the end-to-end tests. This patch
changes that behaviour by making the end-to-end tests tests (including
ec2, lambda, apigatewayv2, iam , ecr, s3 and dynamodb) a requirement for
PRs to be merged.

This change will ensure that code-generator changes are thoroughly
tested before they are merged into the main branch.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>