Make all code-generator pre-submit tests required (#306)
Recently, code-generator PRs were being merged as soon as unit tests
passed, regardless of the outcome of the end-to-end tests. This patch
changes that behavior by making the end-to-end tests (including
ec2, lambda, apigatewayv2, iam , ecr, s3 and dynamodb) a requirement for
PRs to be merged.

This change will ensure that code-generator changes are thoroughly
tested before they are merged into the main branch.
Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.