Prow: Add Karpenter configuration (#308)
Issue #, if available:

Description of changes:
The default Karpenter provisioner in our Prow deployment was not properly configured to make use of the subnets and security groups that are spun up the CDK stack.   Also, the current CDK stack is trying to use both a managed nodegroup and Karpenter, which are not compatible.

This change:
- removes the MNG entirely
- adds some parameters to the `KarpenterAddOn` for instance types, architecture, and capacity type for EC2 resources for node
- points to the subnets and security groups that are spun up by the CDK stack

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.