Fix Prow ingress priority (#303)
Description of changes:
I noticed that the webhooks weren't working. When looking at the webhook panel within the Github app I saw that they were being met with 404s at `/hook`. Doing some digging, I found that the 404s were being returned by the `deck` service rather than `hook`. After adjusting the order that the ingress defines its routes, putting `/hook` and `/metrics` before `/*`, it gave precedence to the more specific subroutes and the webhooks were able to be delivered.

In this PR I also downgraded the `hook` from a `LoadBalancer` to a `ClusterIP` by default. Since this service is exposed through the ingress, which is itself behind a `LoadBalancer`, there is no need to serve it with an ALB. This is how it works for `deck` and so just to stay consistent this is how it should also work for `hook`.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.