Move Karpenter config to GitOps (#313)
Description of changes:
Moves the configuration of the Karpenter provider and provisioner to a GitOps `Kustomization` - so that we aren't restrained by what the EKS blueprints project supports. It also allows us to update the configuration by making changes in Git, without requiring slow CDK deployments.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.