Set Flux and Karpenter to run on Fargate (#314)
Description of changes:
When instantiating a new cluster, Flux and Karpenter have to compete with other resources over the limited initial nodes. This pull request updates the cluster spec to enable pods running in `flux-system` and `karpenter` to use Fargate pods. By doing so, Flux and Karpenter have the resources to set up the provisioner to create additional nodes as needed.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.