Update Github new service controller label (#316)
Description of changes:
With the introduction of the new Github label changes, the documentation generator needs to know about the change from `Service Controller` to `kind/new-service`. This PR moves that string, and some other hardcoded strings, into constants

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.