Remove `v` prefix from `ack-chart` tag jobs (#321)
Description of changes:
The current tag post-submit jobs for the `ack-chart` repository contain `v` prefixes in their definitions. The `ack-chart` repository removes the need for `v` in the tag. This PR removes the `v` from the job requirements.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.