Update yq to `v4.31.2` in all images (#322)
Description of changes:
Updates the version of `yq` in all Prow images to the latest version (`v4.31.2`). `v4.18.1` introduced a new standard syntax for `yq` commands that this change enables in our Prow scripts, and we rely on for the `ack-chart` CD scripts.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.