Upgrade Deploy image Helm to 3.8+ (#323)
Description of changes:
Upgrades the version of Helm in the `deploy` image to 3.8+, so it supports non-experimental support for OCI registries. Also removes the stderr piping in the `ack-chart` CD scripts Helm commands, to make debugging easier for those.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.