Fix CD script bugs for `ack-chart` `update-charts.sh` (#326)
Description of changes:
When migrating the CD scripts from the `ack-chart` repository to `test-infra`, some parts of the `update-chart.sh` script broke when they incorrectly assumed the working directory was still `ack-chart`. This pull request also sets up the proper git username and email, so that commits are correctly associated with the bot account, and adds `find` to the `deploy` image (which is required by the scripts).

This pull request also adds the `ack-chart` repository to the postsubmit job spec, so that `ack-chart` is cloned into the GOPATH of the job. 

These changes were all tested by launching a copy of the script in a rerun of the failed Prow job, and ensuring it passed. [This is the ProwJob](https://prow.ack.aws.dev/log?job=update-ack-chart&id=1633235826118758400) that passed and it produced [this commit in `ack-chart`](https://github.com/aws-controllers-k8s/ack-chart/commit/07b58e6f5a468c990d5c02c59dc98f4f618bc628). [This is the original, failing ProwJob](https://prow.ack.aws.dev/log?job=update-ack-chart&id=1633228367807385600)

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.