Poll ECR public before updating `ack-chart` (#328)
Description of changes:
The `ack-chart` CD process was racing the controller Helm chart release process, and losing most of the time. This caused the `ack-chart` releases to fail and the chart to never be updated. This change introduces a new polling mechanism, with a 5 minute timeout, that ensures the newest version of the Helm chart that started the CD script has been uploaded to ECR public.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.