Add a prow job to validate `metadata.yaml` files (#333)
This patch introduces a new prow job that validates the `metadata.yaml`
files in service controller repositories. It does so by making sure that
the fields are not empty and making sure that the provided URLs are
working properly.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.