Add a new job label for censored Github secrets (#330)
Description of changes:
Successor to https://github.com/aws-controllers-k8s/test-infra/pull/329

This PR creates a new job label `preset-github-secrets` that adds the Github secrets to env vars and a volume mounted into `/etc/github`. It also changes all jobs that use the Github credentials to use the new label. Lastly, it adds `/etc/github` to the list of included directories for the `censor` plugin, so that none of the credentials are leaked.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.