Remove `v` prefix from Helm charts and container images (#341)
Issue #, if available:
https://github.com/aws-controllers-k8s/community/issues/1715

Description of changes:
This pull requests updates all CD scripts that release Helm charts, or rely on the versions defined within Helm charts, to remove or disregard the `v` prefix at the beginning of the Git tag. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.