Use alias in ack-chart `values.yaml` file (#344)
Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/1800

Description of changes:
When adding a new service to the `values.yaml` file of `ack-chart`, don't use `-chart` in its name. The `Chart.yaml` file uses aliases to refer to the services that specifically remove the need for these suffixes.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.