Remove codespell Prow Job (#355)
Description of changes:
Until we fine-tune the codespell configuration in every controller repository, so that we don't get false positives (like fIter -> filter), we should not run the codespell jobs as part of every pull request. It's nice to have spelling checks, but it should not be breaking development.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.