Update logic in ackdiscover's controller.py to fix ECR/GH release version mismatches
We noticed that `gen_services.py` was generating bad version links in its output of the Services page; this was traced back to a recent change in the way image tags are versioned in ECR.  Previously, ECR image tags matched GitHub release strings exactly (e.g., `"1.0.0" == "1.0.0"`), but now the former have a `v` prefix and no longer match (e.g., `"v1.0.0" != "1.0.0"`).

This is an easy fix to rectify the problem by simply prefixing the version string with `v` if that character is not already present.