Update OLM images (#349)
Description of changes:
Updates the OLM images to include the latest changes from https://github.com/aws-controllers-k8s/test-infra/pull/348

Note: I have changed the format of the images to use the short commit SHA as the primary identifier, since it's a more useful indicator of the contents of the images compared to an arbitrary semver. This is the default behaviour if no `VERSION` var is passed to `./push-image`.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.