Improve api-2.json metadata parsers
Originally the script avoided loading all the api-2.json content in
memory and tried to parse until a `},` is observed. Now that sqs team
introduce a new metadata field object, it broke our parsing strategy.

One of the way we can parse this file is assume that the first 100 lines
will have the information we need `serviceFullName` and
`serviceAbbreviation`.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>