Adjut memory/cpu requests and limits for all the prow jobs (#411)
This morning we encountered failures in e2e tests due to etcd timeouts
in the pods running e2e jobs. Investigations revealed that the pod was
scheduled with a memory request for 500Mi, while the test pod was
consuming up to 1.8Gi.

This commit addresses the problem bymaking the following changes:
- Increase the memory requests for all the e2e pods
- Set a memory/cpu limit equal to the requests to prevent contention
  within the executing nodes.

This will ensure that e2e tests have sufficient resources to run without
encountring memory related failures

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.