Calibrate CPU/Memory limits based on the latest observations (#417)
Lately we observed that the release process need much more memory what
was initially calculated. The release pod needs to build images for
multiple architecture and needs up to 4XXXMi in memory.

In addition to that, we've observed that some unit test jobs might need
more memory/cpu.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.