Register more test controllers for ACK `code-generator` PRs (#422)
As the number of supported controllers increase, code-generator changes
are becoming more risky.. We need to increase the number of tested
controllers with the new `code-genetor` changes, to catch as many
regressions as possible.

This patch adds 5 more controllers to the list of controllers that are
tested at each code-generator change. Those controllers are known to
have some edge cases and leverage specific code-generator features that
are not widely used.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.