Setup infra for CARM tests (#423)
This patch adds utilities to help in running e2e tests for CARM (Cross
account resource management). Currently we need a special role and
credentials to allow controllers to manage resources in more than one
account.

This patch mainly adds two things:
- A set of credentials the pytests suites will use to verify that indeed
  resources are being managed in a different account
- A role ARN the controller can assume to manage resources in seperate
  accounts.

Both of the role ARN and crendentials needs to be hidden in the
execution script.

Additionally, since the controllers are, by default, using an assumed
role, we intentionally need to trim the uuid injected in the the
assume-role-session-name, to allow nested assume role. 🤷 ...

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.