Speed up controller auto-generation on code-gen/runtime new releases
In ACK, everytime we release a new runtime/code-generation version, a
prow job will be triggererd to regenerate all the controllers and ship
PRs to their respective github repositories. Historically, we used to
sleep 2 minutes between code-generation, to control the number of e2e
tests/prow jobs running in parallel. The reason was that our cluster had
a static number of nodes, and didn't scale when needed. Now that we use
karpenter to scale our cluster, we can afford faster auto-generation for
the controllers...

Also, the fact that we have 39 controllers to maintain, this job takes
hours to finish, making the process of catching generation errors slow.

This patch, sets the sleep time to 1 second instead.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>