Add new prowjobs to execute e2e tests for runtime PRs
We never run e2e tests for runtime PRs in the past, and generally waited
until the runtime version is pinned in the code-generator repository and
watch the e2e go wild and fireup everything. When things go wrong we end
up making 2 or 3 more patch releases and go throught the same cycle.

Now that we're shipping have features and bug fixes, we no longer can
afford to waste time in the cycle described above. This simply replaces
the runtime of a few controllers and run e2e against them.

Let the show begin.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>